deepseek_local_ai_apps/deepseek_local_docs_analyzer/requirements.txt
openai
streamlit
PyPDF2
python-docx
pandas